Input: goodix - enable reset sequence for Chuwi Hi12 tablet
Goodix touchscreen on Chuwi Hi12 tablet sometimes fails to
initialize properly. When this happens, the driver cannot
read the configuration registers from the device (reads only 0s),
so it does not properly initialize attributes like interrupt flags.

To avoid problems with device power on, the Goodix datasheet
recommends that the driver resets the device. Enabling the reset
sequence for Chuwi Hi12 tablet through dmi quirks.

The full report of this problem can be found in Bugzilla:
https://bugzilla.kernel.org/show_bug.cgi?id=156511

Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>