Input: add documentation for Goodix touchscreen
Goodix touchscreens have some advanced functionality that is
enabled only if the GPIO pins are properly connected and
configured. One of these functionalities is updating the device
configuration.

Add documentation on how to setup the GPIO pins to have access to
full functionality of the device and how to update device
configuration.

Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>